+++ NEWS.md
@@ -18,13 +18,20 @@
 
 ## ver 0.0.1 (GitHub)
 
-* Added the following plotting functions:
-    + `count_null`: return an SQL query that counts the number of NULLs in selected columns
+* Added the following functions:
+    + `count_null`: return an SQL query that counts the number of NULLs in selected columns in a table
+    + `csum_N_pois`: return a DFT-based pmf
+    + `dpmf`: generate a (finite-support) pmf
+    + `impute_em`: impute missing values by the EM algorithm assuming normality of observations
+    + `npmap(...)`: `np.array(list(map(...)))`
     + `plot_lm`: draw linear regression plots
     + `plot_qq`: draw a Normal Q-Q plot
     + `plot_rf`: draw a Residuals vs. Fitted plot
     + `plot_rlev`: draw a Residuals vs. Leverage plot
     + `plot_sl`: draw a Scale-Location plot
+    + `rpmf`: sample from a desired (finite-support) pmf
+    + `simulate_nan(X, nan_rate)` (not made available): randomly replace some values of X with `np.nan`
+
 
 ## ver 0.0.1 (PyPI)